Group by search

+++ backend/src/main/java/com/devsuperior/sds4/dsvendas/dto/SaleDTO.java
@@ -18,8 +18,8 @@ public class SaleDTO  implements Serializable {
     private static final long serialVersionUID = 1L;
 
     private Long id;
-    private Integer visited;
-    private Integer deals;
+    private Long visited;
+    private Long deals;
     private Double amount;
     private LocalDate date;
     private SellerDTO seller;
